PK decided to sync his modifications with the repository and takes a mental note to do this more often ^_^

+++ src/common/entity/characterstats.cpp
@@ -29,11 +29,11 @@ void CharacterStats::sendAllStats(Connection* conn)
   statlist_msg.setStatCount((char)entries.getCount());
   for (size_t i=0; i<entries.getCount(); i++)
   {
-    statlist_msg.setStatId(int(i),entries.get(i)->stat_id);
+    statlist_msg.setStatId(int(i),short(entries.get(i)->stat_id));
     Stat* stat = Server::getServer()->getStatManager()->findById(entries.get(i)->stat_id);
     assert(stat);
     statlist_msg.setName(int(i),stat->getName());
-    statlist_msg.setStatLevel(int(i),entries.get(i)->level);
+    statlist_msg.setStatLevel(int(i),short(entries.get(i)->level));
   }
   ByteStream bs2;
   statlist_msg.serialise(&bs2);
